Rename fields and classes
