+++ emotechat-1.8/src/main/java/de/emotechat/addon/adapter/v18/V18ChatWidthCalculator.java
@@ -1,46 +1,47 @@
-package de.emotechat.addon.adapter.v112;
+package de.emotechat.addon.adapter.v18;
 
 import com.google.common.collect.Lists;
-import com.mojang.realmsclient.gui.ChatFormatting;
+import de.emotechat.addon.gui.chat.render.ChatWidthCalculator;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextComponentString;
+import net.minecraft.util.ChatComponentText;
+import net.minecraft.util.EnumChatFormatting;
+import net.minecraft.util.IChatComponent;
 
 import java.util.List;
 
-public class ChatWidthCalculator extends de.emotechat.addon.gui.chat.render.ChatWidthCalculator {
+public class V18ChatWidthCalculator extends ChatWidthCalculator {
 
     public static String format(String text, boolean useColors) {
-        return !useColors && !Minecraft.getMinecraft().gameSettings.chatColours ? ChatFormatting.stripFormatting(text) : text;
+        return !useColors && !Minecraft.getMinecraft().gameSettings.chatColours ? EnumChatFormatting.getTextWithoutFormattingCodes(text) : text;
     }
 
-    public static List<ITextComponent> calculateLines(ITextComponent component, int chatWidth, FontRenderer renderer, boolean unknown, boolean useColors) {
+    public static List<IChatComponent> calculateLines(IChatComponent component, int chatWidth, FontRenderer renderer, boolean unknown, boolean useColors) {
         int fullLength = 0;
-        ITextComponent fullComponent = new TextComponentString("");
-        List<ITextComponent> output = Lists.newArrayList();
-        List<ITextComponent> input = Lists.newArrayList(component);
+        IChatComponent fullComponent = new ChatComponentText("");
+        List<IChatComponent> output = Lists.newArrayList();
+        List<IChatComponent> input = Lists.newArrayList(component);
 
         for (int j = 0; j < input.size(); ++j) {
-            ITextComponent currentComponent = input.get(j);
-            String inputText = currentComponent.getUnformattedText();
+            IChatComponent currentComponent = input.get(j);
+            String inputText = currentComponent.getUnformattedTextForChat();
             boolean flag = false;
             int newLine = inputText.indexOf('\n');
             if (newLine != -1) {
                 String nextText = inputText.substring(newLine + 1);
                 inputText = inputText.substring(0, newLine + 1);
-                TextComponentString chatcomponenttext = new TextComponentString(nextText);
-                chatcomponenttext.setStyle(currentComponent.getStyle().createShallowCopy());
+                ChatComponentText chatcomponenttext = new ChatComponentText(nextText);
+                chatcomponenttext.setChatStyle(currentComponent.getChatStyle().createShallowCopy());
                 input.add(j + 1, chatcomponenttext);
                 flag = true;
             }
 
-            String formattedText = format(currentComponent.getStyle().getFormattingCode() + inputText, useColors);
+            String formattedText = format(currentComponent.getChatStyle().getFormattingCode() + inputText, useColors);
             String finalText = formattedText.endsWith("\n") ? formattedText.substring(0, formattedText.length() - 1) : formattedText;
             int currentWidth = getStringWidth(renderer, finalText);
 
-            TextComponentString finalComponent = new TextComponentString(finalText);
-            finalComponent.setStyle(currentComponent.getStyle().createShallowCopy());
+            ChatComponentText finalComponent = new ChatComponentText(finalText);
+            finalComponent.setChatStyle(currentComponent.getChatStyle().createShallowCopy());
 
             if (fullLength + currentWidth > chatWidth) {
                 String trimmedText = trimStringToWidth(renderer, formattedText, chatWidth - fullLength, false);
@@ -60,14 +61,14 @@ public class ChatWidthCalculator extends de.emotechat.addon.gui.chat.render.Chat
                     }
 
                     overflowText = FontRenderer.getFormatFromString(trimmedText) + overflowText;
-                    TextComponentString overflowComponent = new TextComponentString(overflowText);
-                    overflowComponent.setStyle(currentComponent.getStyle().createShallowCopy());
+                    ChatComponentText overflowComponent = new ChatComponentText(overflowText);
+                    overflowComponent.setChatStyle(currentComponent.getChatStyle().createShallowCopy());
                     input.add(j + 1, overflowComponent);
                 }
 
                 currentWidth = getStringWidth(renderer, trimmedText);
-                finalComponent = new TextComponentString(trimmedText);
-                finalComponent.setStyle(currentComponent.getStyle().createShallowCopy());
+                finalComponent = new ChatComponentText(trimmedText);
+                finalComponent.setChatStyle(currentComponent.getChatStyle().createShallowCopy());
                 flag = true;
             }
 
@@ -81,7 +82,7 @@ public class ChatWidthCalculator extends de.emotechat.addon.gui.chat.render.Chat
             if (flag) {
                 output.add(fullComponent);
                 fullLength = 0;
-                fullComponent = new TextComponentString("");
+                fullComponent = new ChatComponentText("");
             }
         }
