+++ emotechat-1.8/src/main/java/de/emotechat/addon/adapter/v18/V18EmoteChatAdapter.java
@@ -48,7 +48,6 @@ public class V18EmoteChatAdapter implements EmoteChatAdapter {
         if (packet instanceof C01PacketChatMessage) {
             ((C01PacketChatMessage) packet).getMessage();
         }
-
         return null;
     }
